wow break all the things
